refactor(playground): consolidate agent-builder-only files into domains/agent-builder
Move files whose only consumers are the agent-builder pages and components
out of domains/agents and domains/builder and into domains/agent-builder so
the boundaries match how the code is actually used:

- agent-list: favorite-button (+test) and use-stored-agent-favorite
- skill-list: skill-favorite-button, builder-add-skill-dialog, copy-skill-dialog,
  and use-stored-skill-favorite
- hooks: use-builder-registries, use-copy-skill, use-delete-skill,
  use-stored-skill, use-infrastructure-status
- utils: skill-origin

Delete the dead domains/builder.ts shim (which silently shadowed
domains/builder/index.ts via TS module resolution) and the byte-identical
agent-builder copy of use-builder-settings.ts; rewire the internal callers
to import from @/domains/builder/hooks/use-builder-settings.

Also fix pre-existing API drift where useStoredSkills was called with a
spurious first 'params' argument despite its single-argument signature;
normalize the 6 call sites and remove the now-unused list-params memos.

No behavior change. Shared CMS / LLM / agents code (SkillEditDialog,
SkillChatComposer, use-skill-builder-tool, use-builder-settings,
use-builder-filtered-models, isModelNotAllowedError) stays in place.
