+++ packages/playground/src/domains/agent-builder/components/agent-edit/__tests__/agent-builder-mobile-menu.test.tsx
@@ -43,10 +43,7 @@ const FormHarness = ({ defaultVisibility = 'private', children }: FormHarnessPro
 
 const openDropdown = async () => {
   const trigger = await screen.findByTestId('agent-builder-mobile-menu-trigger');
-  await act(async () => {
-    fireEvent.pointerDown(trigger, { pointerType: 'mouse', button: 0 });
-    fireEvent.click(trigger);
-  });
+  fireEvent.click(trigger);
   await screen.findByRole('menu');
 };
