+++ packages/playground/src/domains/agent-builder/components/skill-list/builder-add-skill-dialog.tsx
@@ -22,7 +22,7 @@ import {
   useInstallBuilderRegistrySkill,
   usePopularBuilderRegistrySkills,
   useSearchBuilderRegistry,
-} from '@/domains/agents/hooks/use-builder-registries';
+} from '@/domains/agent-builder/hooks/use-builder-registries';
 
 // =============================================================================
 // Helpers
