+++ packages/playground/src/domains/agent-builder/components/skill-list/copy-skill-dialog.tsx
