+++ packages/playground/src/domains/agent-builder/contexts/stream-chat-provider.tsx
@@ -91,10 +91,8 @@ export const StreamChatProvider = ({
     send(initialUserMessage);
   }, [initialUserMessage, initialMessages, send]);
 
-  const runningValue = useMemo<RunningContextValue>(
-    () => ({ isRunning: debounceTime === 0 ? isRunning : debouncedIsRunning }),
-    [debouncedIsRunning, isRunning, debounceTime],
-  );
+  const effectiveIsRunning = debounceTime === 0 ? isRunning : debouncedIsRunning;
+  const runningValue = useMemo<RunningContextValue>(() => ({ isRunning: effectiveIsRunning }), [effectiveIsRunning]);
   const messagesValue = useMemo<MessagesContextValue>(() => ({ messages }), [messages]);
   const sendValue = useMemo<SendContextValue>(() => ({ send }), [send]);
