+++ packages/playground/src/domains/agent-builder/hooks/use-builder-registries.ts
