+++ packages/playground/src/domains/agent-builder/hooks/use-delete-skill.ts
