+++ packages/playground/src/domains/agent-builder/hooks/use-save-agent.ts
@@ -4,9 +4,9 @@ import { useCallback } from 'react';
 import type { AgentBuilderEditFormValues } from '../schemas';
 import { formValuesToSaveParams } from '../services/form-values-to-save-params';
 import type { AgentTool } from '../types/agent-tool';
+import { isModelNotAllowedError } from '../utils/is-model-not-allowed';
 import { useStoredAgentMutations } from '@/domains/agents/hooks/use-stored-agents';
 import { useDefaultVisibility } from '@/domains/auth/hooks/use-default-visibility';
-import { isModelNotAllowedError } from '@/domains/builder';
 
 interface UseSaveAgentArgs {
   agentId: string;
