+++ packages/playground/src/domains/agent-builder/hooks/use-stored-skill-favorite.ts
