+++ packages/playground/src/domains/agent-builder/hooks/use-stored-skill.ts
