+++ packages/playground/src/domains/builder/index.ts
@@ -6,6 +6,5 @@ export {
 } from './hooks/use-builder-settings';
 export type { BuilderPickerVisibility } from './hooks/use-builder-settings';
 export { useBuilderFilteredProviders, useBuilderFilteredModels } from './hooks/use-builder-filtered-models';
-export { useInfrastructureStatus } from './hooks/use-infrastructure-status';
 export { isModelNotAllowedError } from './utils/is-model-not-allowed';
 export type { ModelNotAllowedDetails } from './utils/is-model-not-allowed';
