+++ packages/playground/src/pages/agent-builder/favorite/index.tsx
@@ -1,4 +1,4 @@
-import type { ListStoredAgentsParams, ListStoredSkillsParams, StoredSkillResponse } from '@mastra/client-js';
+import type { ListStoredAgentsParams, StoredSkillResponse } from '@mastra/client-js';
 import {
   EmptyState,
   ErrorState,
@@ -49,20 +49,12 @@ export default function AgentBuilderFavoritePage() {
     [],
   );
 
-  const skillListParams = useMemo<ListStoredSkillsParams>(
-    () => ({
-      favoritedOnly: true,
-      orderBy: { field: 'updatedAt', direction: 'DESC' },
-    }),
-    [],
-  );
-
   const { data: agentsData, isLoading: agentsLoading, error: agentsError } = useStoredAgents(agentListParams);
   const {
     data: skillsData,
     isLoading: skillsLoading,
     error: skillsError,
-  } = useStoredSkills(skillListParams, { enabled: tab === 'skills' && features.skills });
+  } = useStoredSkills({ enabled: tab === 'skills' && features.skills });
 
   const agents = agentsData?.agents ?? [];
   const skills = skillsData?.skills ?? [];
